fix warning

+++ src/SSE64SP_Test.cpp
@@ -1771,7 +1771,7 @@ static void asm_debug_test()
 
 FloatTest::FloatTest()
 {
-	FL_ASSERT( RESULT_MAX <= RESULT_BUFFER_MAX );
+	FL_ASSERT( (unsigned int)RESULT_MAX <= (unsigned int)RESULT_BUFFER_MAX );
 	LoopCount= DEFAULT_LOOP;
 	ClearResult();
